requirements.txt
flask
gunicorn
line-bot-sdk==3.5.0
openai
python-dotenv
python-docx